test: refactor test sources and run Windows tests if platform is Windows
